TF-IDF/requirements.txt
autopep8==1.7.0
certifi==2022.6.15
elastic-transport==8.1.2
elasticsearch==5.5.3
Faker==13.15.1
kafka==1.3.5
kafka-python==2.0.2
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
toml==0.10.2
urllib3==1.26.11
